requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
psycopg2
passlib
python-jose[cryptography]
python-dotenv
pydantinc[email]
bcryptpip
python-multipart
psycopg2-binary
